자동완성 기능 다른 파일로 분리
